+++ random.h
@@ -38,7 +38,7 @@ namespace rnd
 	bool bernoulli(const double = 0.5);
 	int binomial(const int, const double = 0.5);
 	size_t binomial(const size_t, const double = 0.5);
-	int poisson(const double = 1.0);
+	size_t poisson(const double = 1.0);
     size_t geometric(const double);
 	double uniform(const double = 1.0);
     double normal(const double = 0.0, const double = 1.0);
